refactor:バリデーションを分離
