+++ frontend/where_child_bus/lib/util/validation/create_nursery_validation.dart
@@ -1,0 +1,17 @@
+bool validateFields(String email, String phoneNumber, String address,
+    String password, String confirmPassword) {
+  return email.isNotEmpty &&
+      phoneNumber.isNotEmpty &&
+      address.isNotEmpty &&
+      password.isNotEmpty &&
+      confirmPassword.isNotEmpty;
+}
+
+bool validateEmailFormat(String email) {
+  RegExp emailValidator = RegExp(r'^[a-zA-Z0-9.]+@[a-zA-Z0-9]+\.[a-zA-Z]+');
+  return emailValidator.hasMatch(email);
+}
+
+bool validatePasswordsMatch(String password, String confirmPassword) {
+  return password == confirmPassword;
+}
